moved contract hook
