+++ front-end/app/tic-tac-toe/scene/TTTScene.tsx
@@ -17,7 +17,7 @@ import { TTTGameEvents } from "@/components/TTT/TTTGameEvents";
 import { TTTSocketEvents } from "@/components/TTT/TTTSocketEvents";
 import { TTTBot } from "@/components/TTT/TTTBot";
 import { useSocket } from "../hooks/useSocket";
-import useContract from "@/app/useContract";
+import useContract from "@/components/hooks/useContract";
 import { TTTModals } from "@/components/TTT/TTTModals";
 import { useGameState } from "../hooks/useGameState";
