+++ front-end/components/Pong/EndModal.tsx
@@ -7,7 +7,7 @@ import { usePongGameState } from "@/app/pong/hooks/usePongGameState";
 import { usePongSocket } from "@/app/pong/hooks/usePongSocket";
 import { usePongUI } from "@/app/pong/hooks/usePongUI";
 import { useKey } from "../hooks/useKey";
-import useContract, { PlayerScore } from "@/app/useContract";
+import useContract, { PlayerScore } from "@/components/hooks/useContract";
 import { initialPongPlayerState } from "@/app/pong/context/PongSockets";
 import { ChevronDownIcon } from "../icons";
 
@@ -38,7 +38,7 @@ const EndModal = () => {
 		showModal
 	} = usePongUI();
 	const escape = useKey(['Escape'])
-	const { submitGameResultRanked, submitGameResultTournament } = useContract();
+	const { submitGameResultRanked, submitGameResultTournament, getTournament } = useContract();
 
 	// State variables
 	const [showResult, setShowResult] = useState("");
@@ -61,10 +61,15 @@ const EndModal = () => {
 					addr: playerState.players[i].addr, score: 1,
 				})
 			}
-			// if (tournament.id !== -1)
-			// 	await submitGameResultTournament(tournament.id, tournament.index, playerScore);
-			// else
-			// 	await submitGameResultRanked(playerScore);
+			if (tournament.id !== -1) {
+				const lol = getTournament(tournament.id);
+				const finished = (await lol).games[tournament.index].finished
+				if (!finished)
+					await submitGameResultTournament(tournament.id, tournament.index, playerScore);
+			}
+			else
+				await submitGameResultRanked(playerScore);
+			
 		}
 		const status = await wsclient?.updateStatus(false, pongGameState.gameId);
 		updatePongGameState({ ...pongGameState, reset: true, pause: true, gameId: "-1" });
